requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pandas>=2.1.0
mplfinance>=0.12.9
